+++ src/store/modules/employee.js
@@ -9,11 +9,13 @@ const state = {
 
 const actions = {
   async actionGetAllEmployee({ commit }) {
+    
     const params = new URLSearchParams();
     params.append('join', 'department');
     params.append('join', 'area');
     params.append('join', 'position');
     params.append('join', 'shift');
+    params.append('join', 'loan');
     params.append('sort', 'name,ASC');
     const res = await httpCommons.get(apiName, { params: params });
     // console.log(res.data)